Convert integration tests to unit tests
